move Name back to lib. Didn't make much sense being in namespaces and making a name module seems too much
