+++ src/debug.rs
@@ -2,7 +2,6 @@
 use crate::{
     attribute::{Attribute, AttributeValue, Prefix},
     misc::{Misc, MiscState},
-    namespaces::Name,
     processing_instruction::ProcessingInstruction,
     prolog::{
         content_particle::ContentParticle,
@@ -23,7 +22,7 @@ use crate::{
         xmldecl::{Standalone, XmlDecl},
     },
     reference::Reference,
-    Document, Tag,
+    Document, Name, Tag,
 };
 use std::fmt::{self, Formatter};
